# Monorepo: rework deps patching in admin plugin for better Webpack performance.