--- a/plugins/woocommerce/changelog/dev-webpack-resources-copy-tweaks
+++ b/plugins/woocommerce/changelog/dev-webpack-resources-copy-tweaks
@@ -1,0 +1,4 @@
+Significance: patch
+Type: dev
+
+Monorepo: tweak patching dependencies for better Webpack perfromance.