### Submission Review Guidelines:

-   I have followed the [WooCommerce Contributing Guidelines](https://github.com/woocommerce/woocommerce/blob/trunk/.github/CONTRIBUTING.md) and the [WordPress Coding Standards](https://make.wordpress.org/core/handbook/best-practices/coding-standards/).
-   I have checked to ensure there aren't other open [Pull Requests](https://github.com/woocommerce/woocommerce/pulls) for the same update/change.
-   I have reviewed my code for [security best practices](https://developer.wordpress.org/apis/security/).
-   Following the above guidelines will result in quick merges and clear and detailed feedback when appropriate.

<!-- You can erase any parts of this template not applicable to your Pull Request. -->

### Changes proposed in this Pull Request:

Related to https://github.com/woocommerce/woocommerce/issues/49549
The changes extracted from https://github.com/woocommerce/woocommerce/pull/49555

Considerations:
- other projects in monorepo are not using `@wordpress/edit-site` in deps, making the revised approach free of side-effects
- the revised approach is pretty much the same as the original (but more explicit), and with better performance.

Profiling data for woocommerce-admin build (non-representative data shows build time reduction of 28.4 seconds)
```
This branch:

. build:project:bundle:  SMP  ⏱  Plugin
. build:project:bundle: NormalModuleReplacementPlugin took 12.34 secs (<- here!)
. build:project:bundle: CopyPlugin took 0.2 secs
. build:project:bundle: WebpackRemoveEmptyScriptsPlugin took 0.115 secs
. build:project:bundle: MiniCssExtractPlugin took 0.1 secs
. build:project:bundle: ForkTsCheckerWebpackPlugin took 0.087 secs
. build:project:bundle: DependencyExtractionWebpackPlugin took 0.034 secs
. build:project:bundle: AssetDataPlugin took 0.002 secs
. build:project:bundle: CustomTemplatedPathPlugin took 0 secs

Trunk

. build:project:bundle:  SMP  ⏱  Plugins
. build:project:bundle: NormalModuleReplacementPlugin took 40.83 secs (<- here!)
. build:project:bundle: CopyPlugin took 0.214 secs
. build:project:bundle: WebpackRemoveEmptyScriptsPlugin took 0.132 secs
. build:project:bundle: MiniCssExtractPlugin took 0.125 secs
. build:project:bundle: ForkTsCheckerWebpackPlugin took 0.082 secs
. build:project:bundle: DependencyExtractionWebpackPlugin took 0.032 secs
. build:project:bundle: AssetDataPlugin took 0.002 secs
. build:project:bundle: CustomTemplatedPathPlugin took 0 secs
```

### Testing

Same as in https://github.com/woocommerce/woocommerce/pull/40884